:sparkles: :: [#289] LectureDetailSettingFeature / 강의 계열 Popup -> BottomSheet
